+++ features/step_definitions/environment_steps.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 Given(/^I'm currently not in a git repository$/) do
   FileUtils.rm_rf '.git'
 end