+++ src/cmd/hack.go
@@ -11,6 +11,10 @@ import (
 	"github.com/spf13/cobra"
 )
 
+type hackConfig struct {
+	TargetBranch string
+}
+
 var hackCmd = &cobra.Command{
 	Use:   "hack <branch>",
 	Short: "Creates a new feature branch off the main development branch",
@@ -26,8 +30,6 @@ the main branch is synced with its upstream counterpart.
 This can be disabled by toggling the "hack-push-flag" configuration:
 $ git town hack-push-flag false`,
 	Run: func(cmd *cobra.Command, args []string) {
-		git.EnsureIsRepository()
-		prompt.EnsureIsConfigured()
 		steps.Run(steps.RunOptions{
 			CanSkip:              func() bool { return false },
 			Command:              "hack",
@@ -37,8 +39,8 @@ $ git town hack-push-flag false`,
 			IsUndo:               false,
 			SkipMessageGenerator: func() string { return "" },
 			StepListGenerator: func() steps.StepList {
-				targetBranchName := checkHackPreconditions(args)
-				return getHackStepList(targetBranchName)
+				config := getHackConfig(args)
+				return getHackStepList(config)
 			},
 		})
 	},
@@ -46,25 +48,34 @@ $ git town hack-push-flag false`,
 		if len(args) == 0 && !abortFlag && !continueFlag {
 			return errors.New("no branch name provided")
 		}
-		return validateMaxArgs(args, 1)
+		err := validateMaxArgs(args, 1)
+		if err != nil {
+			return err
+		}
+		err = git.ValidateIsRepository()
+		if err != nil {
+			return err
+		}
+		prompt.EnsureIsConfigured()
+		return nil
 	},
 }
 
-func checkHackPreconditions(args []string) string {
-	targetBranchName := args[0]
-	if git.HasRemote("origin") {
+func getHackConfig(args []string) (result hackConfig) {
+	result.TargetBranch = args[0]
+	if git.HasRemote("origin") && !git.IsOffline() {
 		script.Fetch()
 	}
-	git.EnsureDoesNotHaveBranch(targetBranchName)
-	return targetBranchName
+	git.EnsureDoesNotHaveBranch(result.TargetBranch)
+	return
 }
 
-func getHackStepList(targetBranchName string) (result steps.StepList) {
+func getHackStepList(config hackConfig) (result steps.StepList) {
 	mainBranchName := git.GetMainBranch()
 	result.AppendList(steps.GetSyncBranchSteps(mainBranchName))
-	result.Append(steps.CreateAndCheckoutBranchStep{BranchName: targetBranchName, ParentBranchName: mainBranchName})
-	if git.HasRemote("origin") && git.ShouldHackPush() {
-		result.Append(steps.CreateTrackingBranchStep{BranchName: targetBranchName})
+	result.Append(&steps.CreateAndCheckoutBranchStep{BranchName: config.TargetBranch, ParentBranchName: mainBranchName})
+	if git.HasRemote("origin") && git.ShouldHackPush() && !git.IsOffline() {
+		result.Append(&steps.CreateTrackingBranchStep{BranchName: config.TargetBranch})
 	}
 	result.Wrap(steps.WrapOptions{RunInGitRoot: true, StashOpenChanges: true})
 	return