+++ src/cmd/sync.go
@@ -37,8 +37,6 @@ When run on the main branch or a perennial branch
 Additionally, when there is a remote upstream,
 the main branch is synced with its upstream counterpart.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		git.EnsureIsRepository()
-		prompt.EnsureIsConfigured()
 		steps.Run(steps.RunOptions{
 			CanSkip: func() bool {
 				return !(git.IsRebaseInProgress() && git.IsMainBranch(git.GetCurrentBranchName()))
@@ -52,17 +50,26 @@ the main branch is synced with its upstream counterpart.`,
 				return fmt.Sprintf("the sync of the '%s' branch", git.GetCurrentBranchName())
 			},
 			StepListGenerator: func() steps.StepList {
-				return getSyncStepList(checkSyncPreconditions())
+				return getSyncStepList(getSyncConfig())
 			},
 		})
 	},
 	PreRunE: func(cmd *cobra.Command, args []string) error {
-		return validateMaxArgs(args, 0)
+		err := validateMaxArgs(args, 0)
+		if err != nil {
+			return err
+		}
+		err = git.ValidateIsRepository()
+		if err != nil {
+			return err
+		}
+		prompt.EnsureIsConfigured()
+		return nil
 	},
 }
 
-func checkSyncPreconditions() (result syncConfig) {
-	if git.HasRemote("origin") {
+func getSyncConfig() (result syncConfig) {
+	if git.HasRemote("origin") && !git.IsOffline() {
 		script.Fetch()
 	}
 	result.InitialBranch = git.GetCurrentBranchName()
@@ -85,9 +92,9 @@ func getSyncStepList(config syncConfig) (result steps.StepList) {
 	for _, branchName := range config.BranchesToSync {
 		result.AppendList(steps.GetSyncBranchSteps(branchName))
 	}
-	result.Append(steps.CheckoutBranchStep{BranchName: config.InitialBranch})
-	if git.HasRemote("origin") && config.ShouldPushTags {
-		result.Append(steps.PushTagsStep{})
+	result.Append(&steps.CheckoutBranchStep{BranchName: config.InitialBranch})
+	if git.HasRemote("origin") && config.ShouldPushTags && !git.IsOffline() {
+		result.Append(&steps.PushTagsStep{})
 	}
 	result.Wrap(steps.WrapOptions{RunInGitRoot: true, StashOpenChanges: true})
 	return