+++ src/drivers/gitlab.go
@@ -1,24 +1,59 @@
 package drivers
 
 import (
+	"errors"
 	"fmt"
 	"net/url"
+	"strings"
+
+	"github.com/Originate/git-town/src/git"
 )
 
-// GitlabCodeHostingDriver provides tools for working with repositories
-// on Gitlab.
-type GitlabCodeHostingDriver struct{}
+type gitlabCodeHostingDriver struct {
+	originURL  string
+	hostname   string
+	repository string
+}
 
-// GetNewPullRequestURL returns the URL of the page
-// to create a new pull request on Gitlab
-func (driver GitlabCodeHostingDriver) GetNewPullRequestURL(repository string, branch string, parentBranch string) string {
+func (d *gitlabCodeHostingDriver) CanBeUsed() bool {
+	return d.hostname == "gitlab.com" || strings.Contains(d.hostname, "gitlab")
+}
+
+func (d *gitlabCodeHostingDriver) CanMergePullRequest(branch, parentBranch string) (bool, error) {
+	return false, nil
+}
+
+func (d *gitlabCodeHostingDriver) GetNewPullRequestURL(branch, parentBranch string) string {
 	query := url.Values{}
 	query.Add("merge_request[source_branch]", branch)
 	query.Add("merge_request[target_branch]", parentBranch)
-	return fmt.Sprintf("https://gitlab.com/%s/merge_requests/new?%s", repository, query.Encode())
+	return fmt.Sprintf("%s/merge_requests/new?%s", d.GetRepositoryURL(), query.Encode())
 }
 
-// GetRepositoryURL returns the URL of the given repository on Gitlab
-func (driver GitlabCodeHostingDriver) GetRepositoryURL(repository string) string {
-	return "https://gitlab.com/" + repository
+func (d *gitlabCodeHostingDriver) GetRepositoryURL() string {
+	return "https://gitlab.com/" + d.repository
+}
+
+func (d *gitlabCodeHostingDriver) MergePullRequest(options MergePullRequestOptions) (string, error) {
+	return "", errors.New("shipping pull requests via the Gitlab API is currently not supported. If you need this functionality, please vote for it by opening a ticket at https://github.com/originate/git-town/issues")
+}
+
+func (d *gitlabCodeHostingDriver) HostingServiceName() string {
+	return "Gitlab"
+}
+
+func (d *gitlabCodeHostingDriver) SetOriginURL(originURL string) {
+	d.originURL = originURL
+	d.hostname = git.GetURLHostname(originURL)
+	d.repository = git.GetURLRepositoryName(originURL)
+}
+
+func (d *gitlabCodeHostingDriver) GetAPITokenKey() string {
+	return ""
+}
+
+func (d *gitlabCodeHostingDriver) SetAPIToken(apiToken string) {}
+
+func init() {
+	registry.RegisterDriver(&gitlabCodeHostingDriver{})
 }