+++ src/git/environment.go
@@ -1,10 +1,30 @@
 package git
 
-import "github.com/Originate/git-town/src/util"
+import (
+	"errors"
 
-// EnsureIsRepository asserts that the current directory is in a repository
-func EnsureIsRepository() {
-	util.Ensure(IsRepository(), "This is not a Git repository.")
+	"github.com/Originate/git-town/src/util"
+)
+
+// IsOffline returns whether Git Town is currently in offline mode
+func IsOffline() bool {
+	return util.StringToBool(getConfigurationValueWithDefault("git-town.offline", "false"))
+}
+
+// ValidateIsOnline asserts that Git Town is not in offline mode
+func ValidateIsOnline() error {
+	if IsOffline() {
+		return errors.New("This command requires an active internet connection")
+	}
+	return nil
+}
+
+// ValidateIsRepository asserts that the current directory is in a repository
+func ValidateIsRepository() error {
+	if IsRepository() {
+		return nil
+	}
+	return errors.New("This is not a Git repository")
 }
 
 // IsRepository returns whether or not the current directory is in a repository