+++ src/steps/create_pull_request_step.go
@@ -13,10 +13,9 @@ type CreatePullRequestStep struct {
 }
 
 // Run executes this step.
-func (step CreatePullRequestStep) Run() error {
-	driver := drivers.GetCodeHostingDriver()
-	repository := git.GetURLRepositoryName(git.GetRemoteOriginURL())
+func (step *CreatePullRequestStep) Run() error {
+	driver := drivers.GetActiveDriver()
 	parentBranch := git.GetParentBranch(step.BranchName)
-	script.OpenBrowser(driver.GetNewPullRequestURL(repository, step.BranchName, parentBranch))
+	script.OpenBrowser(driver.GetNewPullRequestURL(step.BranchName, parentBranch))
 	return nil
 }