+++ src/steps/squash_merge_branch_step.go
@@ -1,12 +1,9 @@
 package steps
 
 import (
-	"log"
-
 	"github.com/Originate/git-town/src/git"
 	"github.com/Originate/git-town/src/prompt"
 	"github.com/Originate/git-town/src/script"
-	"github.com/Originate/git-town/src/util"
 )
 
 // SquashMergeBranchStep squash merges the branch with the given name into the current branch
@@ -17,27 +14,24 @@ type SquashMergeBranchStep struct {
 }
 
 // CreateAbortStep returns the abort step for this step.
-func (step SquashMergeBranchStep) CreateAbortStep() Step {
-	return DiscardOpenChangesStep{}
+func (step *SquashMergeBranchStep) CreateAbortStep() Step {
+	return &DiscardOpenChangesStep{}
 }
 
 // CreateUndoStepAfterRun returns the undo step for this step after it is run.
-func (step SquashMergeBranchStep) CreateUndoStepAfterRun() Step {
-	return RevertCommitStep{Sha: git.GetCurrentSha()}
+func (step *SquashMergeBranchStep) CreateUndoStepAfterRun() Step {
+	return &RevertCommitStep{Sha: git.GetCurrentSha()}
 }
 
 // GetAutomaticAbortErrorMessage returns the error message to display when this step
 // cause the command to automatically abort.
-func (step SquashMergeBranchStep) GetAutomaticAbortErrorMessage() string {
+func (step *SquashMergeBranchStep) GetAutomaticAbortErrorMessage() string {
 	return "Aborted because commit exited with error"
 }
 
 // Run executes this step.
-func (step SquashMergeBranchStep) Run() error {
-	err := script.RunCommand("git", "merge", "--squash", step.BranchName)
-	if err != nil {
-		log.Fatal("Error squash merging:", err)
-	}
+func (step *SquashMergeBranchStep) Run() error {
+	script.SquashMerge(step.BranchName)
 	commitCmd := []string{"git", "commit"}
 	if step.CommitMessage != "" {
 		commitCmd = append(commitCmd, "-m", step.CommitMessage)
@@ -46,12 +40,12 @@ func (step SquashMergeBranchStep) Run() error {
 	if author != git.GetLocalAuthor() {
 		commitCmd = append(commitCmd, "--author", author)
 	}
-	util.GetCommandOutput("sed", "-i", "-e", "s/^/# /g", ".git/SQUASH_MSG")
+	git.CommentOutDefaultSquashCommitMessage()
 	return script.RunCommand(commitCmd...)
 }
 
 // ShouldAutomaticallyAbortOnError returns whether this step should cause the command to
 // automatically abort if it errors.
-func (step SquashMergeBranchStep) ShouldAutomaticallyAbortOnError() bool {
+func (step *SquashMergeBranchStep) ShouldAutomaticallyAbortOnError() bool {
 	return true
 }